Create authenticated rotue to video collections

+++ src/controllers/video.controller.js
@@ -19,7 +19,7 @@ module.exports = {
         .populate({
           path: "comments",
           select: "commentBody",
-          populate: { path: "userId", select:"name avatar" },
+          populate: { path: "userId", select: "name avatar" },
         });
       res.status(200).json({ message: "Video found", data: video });
     } catch (err) {
@@ -29,7 +29,7 @@ module.exports = {
 
   async create(req, res) {
     try {
-      const userId  = req.user;
+      const userId = req.user;
       const user = await User.findById(userId);
       if (!user) {
         throw new Error("Invalid user");
@@ -44,11 +44,20 @@ module.exports = {
   async update(req, res) {
     try {
       const { videoId } = req.params;
+      const userId = req.user;
 
-      const video = await Video.findByIdAndUpdate(videoId, req.body, {
+      const video = await Video.findById(videoId);
+
+      if (video.userId.toString() !== userId) {
+        res.status(403).json({ message: "unauthorized user" });
+        return;
+      }
+
+      const videoUpdated = await Video.findByIdAndUpdate(videoId, req.body, {
         new: true,
       });
-      res.status(200).json({ message: "video updated", data: video });
+
+      res.status(200).json({ message: "video updated", data: videoUpdated });
     } catch (err) {
       res
         .status(400)
@@ -59,8 +68,17 @@ module.exports = {
   async destroy(req, res) {
     try {
       const { videoId } = req.params;
-      const video = await Video.deleteOne({_id:videoId});
-      res.status(200).json({ message: "video deleted", data: video });
+      const userId = req.user;
+
+      const video = await Video.findById(videoId);
+
+      if (video.userId.toString() !== userId) {
+        res.status(403).json({ message: "unauthorized user" });
+        return;
+      }
+
+      const videoDeleted = await Video.deleteOne({ _id: videoId });
+      res.status(200).json({ message: "video Deleted", data: videoDeleted });
     } catch (err) {
       res
         .status(400)
